requirements.txt
PyPDF2>=3.0.0
python-dotenv>=1.0.0
langchain>=0.0.208
langchain-google-genai>=0.0.1
google-genai>=1.0.0
pinecone-client>=2.2.1
